Move internal _is_unittest_unexpected_success_a_failure to "compat" module
Fix #1815
