+++ _pytest/skipping.py
@@ -216,18 +216,6 @@ def check_strict_xfail(pyfuncitem):
             pytest.fail('[XPASS(strict)] ' + explanation, pytrace=False)
 
 
-def _is_unittest_unexpected_success_a_failure():
-    """Return if the test suite should fail if a @expectedFailure unittest test PASSES.
-
-    From https://docs.python.org/3/library/unittest.html?highlight=unittest#unittest.TestResult.wasSuccessful:
-        Changed in version 3.4: Returns False if there were any
-        unexpectedSuccesses from tests marked with the expectedFailure() decorator.
-
-    TODO: this should be moved to the "compat" module.
-    """
-    return sys.version_info >= (3, 4)
-
-
 @pytest.hookimpl(hookwrapper=True)
 def pytest_runtest_makereport(item, call):
     outcome = yield
@@ -236,6 +224,7 @@ def pytest_runtest_makereport(item, call):
     evalskip = getattr(item, '_evalskip', None)
     # unitttest special case, see setting of _unexpectedsuccess
     if hasattr(item, '_unexpectedsuccess') and rep.when == "call":
+        from _pytest.compat import _is_unittest_unexpected_success_a_failure
         if item._unexpectedsuccess:
             rep.longrepr = "Unexpected success: {0}".format(item._unexpectedsuccess)
         else:
